Fixed line length; Added comments; Moved Mock
